# api/openapi/registration.yaml
openapi: "3.0.0"
info:
  title: "Registration & Onboarding API"
  version: "1.0.0"
  description: |
    Bot-facing endpoints covering the pilot registration and VA onboarding flow.
    Consumed by comrade-bot to register pilots, initialise Discord servers as VAs,
    join memberships, and look up user status.

    All endpoints require authentication via three headers sent by the bot.
servers:
  - url: /api/v1
security:
  - ApiKey: []
    ServerId: []
    DiscordId: []
paths:
...
  /server/init:
    post:
      operationId: initServer
      summary: Bootstrap a Discord server with a VA Code / ID
      description: |
        Creates a minimal VA record linked to the Discord server in X-Discord-Server-Id.
        Requires the calling user to already be registered (/pilots/register) and
        makes that user the VA admin. Discord collects only the VA Code / ID;
        flight matching setup continues in Vizburo.
      tags: [registration]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/InitServerRequest"
      responses:
        "201":
          description: Server bootstrapped and minimal VA created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InitServerResponse"
        "400":
          description: Missing fields or unregistered admin prerequisite
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Missing or invalid X-API-Key authentication
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Missing required Discord context headers X-Discord-User-Id or X-Discord-Server-Id
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: Discord server already registered as a VA, or VA Code / ID already exists
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "422":
          description: Validation failure (missing required fields)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ValidationErrorResponse"
        "500":
          description: Database error during VA creation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  securitySchemes:
    ApiKey:
      type: apiKey
      in: header
      name: X-API-Key
      description: Bot API key issued per VA
    ServerId:
      type: apiKey
      in: header
      name: X-Discord-Server-Id
      description: Discord server (guild) ID — used to resolve the VA context
    DiscordId:
      type: apiKey
      in: header
      name: X-Discord-User-Id
      description: Discord user ID of the caller
  schemas:
...
    ErrorDetail:
      type: object
      required: [code, message]
      properties:
        code:
          type: string
          description: Machine-readable error code (e.g. IFC_USER_NOT_FOUND)
          example: IFC_USER_NOT_FOUND
        message:
          type: string
          description: Human-readable error description
          example: IFC user not found. Please verify your IFC username.
    ValidationFieldError:
      type: object
      required: [field, message]
      properties:
        field:
          type: string
          example: ifc_id
        message:
          type: string
          example: ifc_id is required
    ErrorResponse:
      type: object
      required: [status, responseTimeMs]
      properties:
        status:
          type: string
          enum: [error]
        error:
          $ref: "#/components/schemas/ErrorDetail"
        responseTimeMs:
          type: integer
          format: int64
    ValidationErrorResponse:
      type: object
      required: [status, responseTimeMs]
      properties:
        status:
          type: string
          enum: [error]
        error:
          type: object
          required: [code, message]
          properties:
            code:
              type: string
              enum: [VALIDATION_FAILED]
            message:
              type: string
            fields:
              type: array
              items:
                $ref: "#/components/schemas/ValidationFieldError"
        responseTimeMs:
          type: integer
          format: int64

    # ── /pilots/register ────────────────────────────────────────────────────
...
    InitServerRequest:
      type: object
      required: [va_code]
      properties:
        va_code:
          type: string
          description: Short admin-facing VA Code / ID (e.g. "IFE")
          example: IFE
    InitServerData:
      type: object
      required: [success, message, va_code, setup_required]
      properties:
        success:
          type: boolean
        message:
          type: string
          example: Server initialised successfully
        va_code:
          type: string
          example: IFE
        setup_required:
          type: boolean
          description: True when the VA still needs Vizburo setup before flight matching is operational.
          example: true
        dashboard_url:
          type: string
          description: Optional signed dashboard URL, if generated by the backend.
          example: https://viz.example.com/auth/login?token=...
        setup_url:
          type: string
          description: Optional signed setup URL, if generated by the backend.
          example: https://viz.example.com/auth/login?token=...
    InitServerResponse:
      type: object
      required: [status, responseTimeMs]
      properties:
        status:
          type: string
          enum: [ok]
        result:
          $ref: "#/components/schemas/InitServerData"
        responseTimeMs:
          type: integer
          format: int64

    # ── /memberships/join ───────────────────────────────────────────────────
...